data-ingestion-coinbase-market-data/requirements.txt
websockets
kafka-python
python-dotenv